requirements.txt
streamlit==1.34.0
opencv-python==4.9.0.80
numpy==1.23.5
torch==1.13.1+cu116
reportlab==4.2.0
pillow==10.3.0
matplotlib==3.8.4
PyYAML==6.0.1
requests==2.31.0
scipy==1.13.0
tqdm==4.66.4
protobuf==3.20.0
tensorboard==2.16.2
pandas==2.2.2
seaborn==0.13.2
ipython==8.18.1
psutil==5.9.8
thop==0.1.1.post2209072238

# Optional dependencies
# coremltools>=4.1  # CoreML export
# onnx>=1.9.0  # ONNX export
# onnx-simplifier>=0.3.6  # ONNX simplifier
# scikit-learn==0.19.2  # CoreML quantization
# tensorflow>=2.4.1  # TFLite export
# tensorflowjs>=3.9.0  # TF.js export
# openvino-dev  # OpenVINO export
# albumentations>=1.0.3
# pycocotools>=2.0  # COCO mAP
# roboflow
